+++ src/Contracts/Database.php
@@ -16,5 +16,5 @@ interface Database
         int $start_id = 1
     );
 
-    public function search(): int;
+    public function find(): int;
 }